// src/app/core/http/model/message.ts
/**
 * 1HEED APi
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0-oas3
 * Contact: you@your-company.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */import { Attachments } from './attachments';
import { FacebookUser } from './facebookUser';
import { Shares } from './shares';
import { Tags } from './tags';
import { UserList } from './userList';


export interface Message { 
    id?: string;
    message?: string;
    sticker?: string;
    from?: FacebookUser;
    to?: UserList;
    tags?: Tags;
    shares?: Shares;
    attachments?: Attachments;
    createdTime?: string;
}